subsys: usb: device_next: add new MIDI 2.0 device class
This adds a new USB device class (based on usb/device_next) that implements
revision 2.0 of the MIDIStreaming interface, a sub-class of the USB audio
device class. In practice, the MIDI interface is much more simple and has
little in common with Audio, so it makes sense to have it as a separate class
driver.

MIDI inputs and outputs are configured through the device tree, under a node
`compatible = "zephyr,usb-midi"`. As per the USB-MIDI2.0 specification,
a single usb-midi interface can convey up to 16 Universal MIDI groups,
comprising 16 channels each. Data is carried from/to the host via so-called
Group Terminals, that are organized in Group Terminal Blocks. They are
represented as children of the usb-midi interface in the device tree.

From the Zephyr application programmer perspective, MIDI data is exchanged with
the host through the device associated with the `zephyr,usb-midi` interface,
using the following API:

* Send a Universal MIDI Packet to the host: `usb_midi_send(device, pkt)`
* Universal MIDI Packets from the host are delivered to the function passed in
  `usb_midi_set_callback(device, void (device, pkt){...})`

Compliant USB-MIDI 2.0 devices are required to expose a USB-MIDI1.0 interface
as alt setting 0, and the 2.0 interface on alt setting 1. To avoid the extra
complexity of generating backward compatible USB descriptors and translating
Universal MIDI Packets from/to the old USB-MIDI1.0 format, this driver
generates an empty MIDI1.0 interface (no input/output); and therefore will only
be able to exchange MIDI data when the host has explicitely enabled MIDI2.0
(alt setting 1)

This implementation is based on the following documents, which are referred to
in the inline comments:

* `midi20`: [Universal Serial Bus Device Class Definition for MIDI Devices Release 2.0](https://www.usb.org/sites/default/files/USB%20MIDI%20v2_0.pdf)
* `ump112`: [Universal MIDI Packet (UMP) Format and MIDI 2.0 Protocol With MIDI 1.0 Protocol in UMP Format _Document Version 1.1.2_](https://midi.org/universal-midi-packet-ump-and-midi-2-0-protocol-specification)

Signed-off-by: Titouan Christophe <moiandme@gmail.com>