display: stm32_ltdc: use pixel from device tree instead of Kconfig
The device tree binding lcd-controller, which is included in the
stm32-ltdc binding, requires a mandatory pixel-format property.
However, the stm32-ltdc driver does not use it at all, and uses
a custom KConfig option instead.

This is confusing, because the dt property must be specified but has
no consequence. Moreover, the Kconfig option is a choice which
generally clutters the boards configuration.

Let's therefore make use of the device tree property directly in the
stm32 ltdc driver.

Signed-off-by: Titouan Christophe <titouan.christophe@mind.be>