drivers: audio: add new driver for Cirrus CS47L63
Add a new driver for the CS47L63 audio codec and DSP. This initial
implementation only supports audio output on the unique mono headphones
out present on this chip. This driver currently assumes that digital
audio is delivered to the CS47L63 through its 1st audio serial port (ASP1).

Also, move its dts binding out of sound/ (where it is the only entry),
so that it sits together with other audio codec drivers in audio/

Signed-off-by: Titouan Christophe <titouan.christophe@mind.be>